feat(search): cap /search to the realm of the request
The search page served every realm the same results. The city filter listed
every municipality in the database, and a search with no city filter defaulted
to all of them, so opencouncil.fr answered with Greek subjects.

search() now resolves the realm and caps the query to it. An absent city filter
defaults to the realm's municipalities. An explicit filter keeps only the ids
inside the realm, so a city of another realm narrows the search to nothing. An
empty result of that cap returns early, because an empty list means "no city
filter" to buildFilters.

The cap applies to city ids, not to the index's city_realm field. Documents
indexed before that field existed carry no value for it, so a term filter on it
drops them until PGSync re-syncs the whole index.

The AI filter extraction reads a realm-scoped city list, and its extracted ids
pass the same cap. Its guess is advisory, so an id outside the realm drops out
and the search keeps the ids resolved before it.

buildSearchQuery merged the extracted city ids a second time, over the ids the
caller passed. That put the raw ids back, past the cap. The caller already
merges both the city ids and the date range.

The realm travels to search() as an argument, defaulting to getRealm(). The MCP
tool handlers may run outside a request scope, where the Host cannot be read, so
they pass it explicitly.
