+++ src/lib/search/index.ts
@@ -1,17 +1,8 @@
 "use server";
 
-import { Client } from '@elastic/elasticsearch';
-import { Prisma } from '@prisma/client';
-import prisma from "@/lib/db/prisma";
-import { SearchRequest, SearchResponse, SearchResultLight, SearchResultDetailed, SubjectDocument, ExtractedFilters } from './types';
-import { buildSearchQuery } from './query';
-import { extractFilters, processFilters, NO_EXTRACTED_FILTERS } from './filters';
-import { sendErrorAdminAlert } from '@/lib/discord';
-import { executeElasticsearchWithRetry } from './retry';
-import { partitionHits, reportOrphanedHits } from './hits';
-import { getCities } from '@/lib/db/cities';
-import { logSearchQuery } from '@/lib/db/searchQueries';
-import { env } from '@/env.mjs';
+import { SearchRequest, SearchResponse } from './types';
+import { searchInRealm } from './core';
+import { getRealm } from '@/lib/realm.server';
 
 // Re-export types
 export type {
@@ -20,357 +11,25 @@ export type {
     SearchConfig,
 } from './types';
 
-// Initialize Elasticsearch client
-const client = new Client({
-    node: env.ELASTICSEARCH_URL,
-    auth: {
-        apiKey: env.ELASTICSEARCH_API_KEY
-    }
-});
-
-// Helper function for essential logs that should always be shown
-const logEssential = (message: string, data?: any) => {
-    console.log(`[Search Analytics] ${message}`, data || '');
-};
-
-// Relations needed to turn a speaker segment into a detailed search result entry
-const subjectDiscussionSegmentInclude = {
-    meeting: {
-        include: {
-            city: true
-        }
-    },
-    speakerTag: {
-        include: {
-            person: {
-                include: {
-                    roles: {
-                        include: {
-                            party: true,
-                            city: true,
-                            administrativeBody: true
-                        }
-                    }
-                }
-            }
-        }
-    },
-    utterances: true,
-    summary: true
-} satisfies Prisma.SpeakerSegmentInclude;
-
-type SubjectDiscussionSegment = Prisma.SpeakerSegmentGetPayload<{ include: typeof subjectDiscussionSegmentInclude }>;
-
+/**
+ * Search, capped to the realm of the incoming request.
+ *
+ * This module is a Server Action boundary, so every argument here is
+ * caller-controlled — including anything a browser cares to send. The realm is
+ * therefore resolved from the request Host and is not part of `SearchRequest`;
+ * a client cannot widen the search to another tenant by asking for one.
+ *
+ * Server-side callers that carry their own realm context and run outside a
+ * request scope (the MCP tool handlers) call `searchInRealm` in ./core
+ * directly, rather than reaching this entry point.
+ *
+ * `getRealm` is handed over as a resolver rather than awaited here, so a
+ * failure reading the request headers lands in the same logged, sanitized
+ * error path as every other search failure.
+ */
 export async function search(
     request: SearchRequest,
     options?: { skipQueryLog?: boolean }
 ): Promise<SearchResponse> {
-    try {
-        // Persist the query for usage analytics. Skipped for paginated requests
-        // (same query, next page), filter-only searches and internal callers.
-        const queryText = request.query?.trim() ?? '';
-        if (!options?.skipQueryLog && queryText && (request.config?.from ?? 0) === 0) {
-            void logSearchQuery(queryText);
-        }
-
-        // Get default city IDs if none provided
-        let cityIds = request.cityIds;
-        if (!cityIds || cityIds.length === 0) {
-            const cities = await getCities();
-            cityIds = cities.map(city => city.id);
-        }
-
-        // Log search session start with query and filters
-        logEssential('Search Session Started', {
-            query: request.query,
-            filters: {
-                cityIds,
-                personIds: request.personIds,
-                partyIds: request.partyIds,
-                topicIds: request.topicIds,
-                dateRange: request.dateRange,
-                hasLocations: request.locations ? request.locations.length > 0 : false
-            }
-        });
-
-        // Extract filters from the query using AI (non-fatal — search works without it)
-        let extractedFilters: ExtractedFilters = NO_EXTRACTED_FILTERS;
-        if (queryText) {
-            try {
-                extractedFilters = await extractFilters(queryText);
-                logEssential('[Search] Extracted filters:', extractedFilters);
-            } catch (error) {
-                console.error('[Search] AI filter extraction failed, continuing without AI filters:', error);
-            }
-        }
-
-        // Process filters and resolve locations (non-fatal)
-        let processedFilters: Awaited<ReturnType<typeof processFilters>> = {
-            cityIds: undefined,
-            dateRange: undefined,
-            locations: undefined,
-        };
-        try {
-            processedFilters = await processFilters(extractedFilters);
-        } catch (error) {
-            console.error('[Search] Filter processing failed, continuing without processed filters:', error);
-        }
-
-        // Merge with explicit filters
-        const mergedRequest: SearchRequest = {
-            ...request,
-            cityIds: processedFilters.cityIds || cityIds,
-            dateRange: processedFilters.dateRange || request.dateRange,
-            locations: processedFilters.locations || request.locations
-        };
-
-        // Build and execute the search query with retry logic
-        const searchQuery = buildSearchQuery(mergedRequest, extractedFilters);
-        
-        logEssential('Executing search query', { 
-            hasSemanticSearch: request.config?.enableSemanticSearch 
-        });
-
-        const response = await executeElasticsearchWithRetry(
-            () => client.search<SubjectDocument>(searchQuery),
-            'Search'
-        );
-
-        // Get total hits
-        const total = response.hits.total as { value: number; relation: string };
-        const totalHits = total.value;
-
-        // Log search session completion with results summary
-        logEssential('Search Session Completed', {
-            query: request.query,
-            results: {
-                totalHits,
-                resultCount: response.hits.hits.length,
-                took: `${response.took}ms`,
-                topScore: response.hits.hits[0]?._score || 0
-            }
-        });
-
-        // Process the results
-        const subjectIds = response.hits.hits
-            .map(hit => hit._source?.id)
-            .filter((id): id is string => id !== undefined);
-
-        // Fetch all subjects in a single query
-        const subjects = await prisma.subject.findMany({
-            where: { id: { in: subjectIds } },
-            include: {
-                location: true,
-                topic: true,
-                councilMeeting: {
-                    include: {
-                        city: true,
-                        administrativeBody: true
-                    }
-                },
-                introducedBy: {
-                    include: {
-                        roles: {
-                            include: {
-                                party: true,
-                                city: true,
-                                administrativeBody: true
-                            }
-                        }
-                    }
-                },
-                contributions: {
-                    include: {
-                        speaker: {
-                            include: {
-                                roles: {
-                                    include: {
-                                        party: true,
-                                        city: true,
-                                        administrativeBody: true
-                                    }
-                                }
-                            }
-                        }
-                    }
-                },
-                highlights: true,
-                decision: true,
-                discussedIn: {
-                    include: {
-                        topic: true
-                    }
-                }
-            }
-        });
-
-        // Create a map of subjects by ID for efficient lookup
-        const subjectMap = new Map(subjects.map(subject => [subject.id, subject]));
-
-        // For detailed results, fetch the speaker segments discussing each subject
-        // (segments whose utterances are tagged with the subject via discussionSubjectId)
-        const segmentsBySubject = new Map<string, SubjectDiscussionSegment[]>();
-        if (request.config?.detailed && subjectIds.length > 0) {
-            const subjectIdSet = new Set(subjectIds);
-            const segments = await prisma.speakerSegment.findMany({
-                where: {
-                    utterances: { some: { discussionSubjectId: { in: subjectIds } } }
-                },
-                include: subjectDiscussionSegmentInclude,
-                orderBy: { startTimestamp: 'asc' }
-            });
-            for (const segment of segments) {
-                const discussedSubjectIds = new Set(
-                    segment.utterances
-                        .map(u => u.discussionSubjectId)
-                        .filter((id): id is string => id !== null && subjectIdSet.has(id))
-                );
-                for (const discussedSubjectId of discussedSubjectIds) {
-                    const list = segmentsBySubject.get(discussedSubjectId);
-                    if (list) {
-                        list.push(segment);
-                    } else {
-                        segmentsBySubject.set(discussedSubjectId, [segment]);
-                    }
-                }
-            }
-        }
-
-        // Get all location IDs for coordinates query
-        const locationIds = subjects
-            .map(subject => subject.location?.id)
-            .filter((id): id is string => id !== undefined);
-
-        // Fetch all location coordinates in a single query
-        const locationCoordinates = await prisma.$queryRaw<Array<{ id: string; x: number; y: number }>>`
-            SELECT id, ST_X(coordinates::geometry) as x, ST_Y(coordinates::geometry) as y
-            FROM "Location"
-            WHERE id = ANY(${locationIds})
-            AND type = 'point'
-        `;
-
-        // Create a map of location coordinates by ID
-        const locationCoordinatesMap = new Map(
-            locationCoordinates.map(loc => [loc.id, { x: loc.x, y: loc.y }])
-        );
-
-        // The ES query filters on the indexed released flag, but the database
-        // is the source of truth: a meeting unreleased after indexing (PGSync
-        // lag or outage) must not surface. Re-checked here, at the hydration
-        // boundary, so every search consumer gets the defense and the drop
-        // counts below see it.
-        const { resolved, orphanedIds, unreleasedIds, droppedWithoutSource } = partitionHits(
-            response.hits.hits,
-            subjectMap,
-            subject => subject.councilMeeting.released,
-        );
-
-        if (orphanedIds.length > 0 || unreleasedIds.length > 0 || droppedWithoutSource > 0) {
-            logEssential('[Search] Dropped unresolvable hits', { orphanedIds, unreleasedIds, droppedWithoutSource });
-            void reportOrphanedHits({
-                orphanedIds,
-                unreleasedIds,
-                droppedWithoutSource,
-                // Filter-only searches have no query text; label them so the
-                // alert reads sensibly instead of showing an empty string.
-                query: queryText || '(filter-only)',
-                index: env.ELASTICSEARCH_INDEX,
-            });
-        }
-
-        const results = await Promise.all(
-            resolved.map(async ({ hit, subject }) => {
-                // Get location coordinates if available
-                let locationWithCoordinates = null;
-                if (subject.location) {
-                    const coordinates = locationCoordinatesMap.get(subject.location.id);
-                    if (coordinates) {
-                        locationWithCoordinates = {
-                            ...subject.location,
-                            coordinates
-                        };
-                    }
-                }
-
-                // Base result with common fields
-                const baseResult: SearchResultLight = {
-                    ...subject,
-                    location: locationWithCoordinates,
-                    score: hit._score || 0,
-                    councilMeeting: subject.councilMeeting,
-                    votes: [],
-                    attendance: []
-                };
-
-                // If detailed results are requested, add speaker segment text
-                if (request.config?.detailed) {
-                    const speakerSegments = (segmentsBySubject.get(subject.id) ?? [])
-                        .filter(segment => {
-                            const text = segment.utterances.map(u => u.text).join(' ');
-                            const hasPerson = segment.speakerTag?.person != null;
-                            const hasRoles = Array.isArray(segment.speakerTag?.person?.roles);
-                            return text.length >= 100 && hasPerson && hasRoles;
-                        })
-                        .map(segment => ({
-                            id: segment.id,
-                            startTimestamp: segment.startTimestamp,
-                            endTimestamp: segment.endTimestamp,
-                            meeting: segment.meeting,
-                            person: segment.speakerTag?.person || null,
-                            text: segment.utterances.map(u => u.text).join(' '),
-                            summary: segment.summary ? { text: segment.summary.text } : null
-                        }));
-
-                    return {
-                        ...baseResult,
-                        speakerSegments,
-                        context: subject.context
-                    } as SearchResultDetailed;
-                }
-
-                return baseResult;
-            })
-        );
-
-        // ES's total includes hits we dropped; subtract this page's drops so the
-        // count degrades along with the results. Still approximate — other pages
-        // may hold more drops, so callers that must not leak the existence of
-        // hidden content should withhold the total whenever `dropped` > 0.
-        const dropped = response.hits.hits.length - resolved.length;
-        return {
-            results,
-            total: totalHits - dropped,
-            dropped,
-        };
-    } catch (error) {
-        const errorMessage = error instanceof Error ? error.message : 'Unknown error';
-
-        // Log search session failure
-        logEssential('Search Session Failed', {
-            query: request.query,
-            error: errorMessage,
-            filters: {
-                cityIds: request.cityIds,
-                personIds: request.personIds,
-                partyIds: request.partyIds,
-                topicIds: request.topicIds,
-                dateRange: request.dateRange,
-                hasLocations: request.locations ? request.locations.length > 0 : false
-            }
-        });
-
-        // Notify team via Discord (fire-and-forget)
-        sendErrorAdminAlert({
-            source: 'Search',
-            error: errorMessage,
-            context: {
-                query: request.query,
-                cityIds: request.cityIds?.join(', '),
-                personIds: request.personIds?.join(', '),
-                partyIds: request.partyIds?.join(', '),
-            },
-        }).catch(() => {});
-
-        throw new Error('Failed to execute search');
-    }
+    return searchInRealm(request, getRealm, options);
 }
